feat: extend emoji picker
